+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+**Version 0.9.0** - October 2025<br/>
+
+-   NEW: Data export functionality (exportDaily, exportAll commands)<br/>
+-   FIX: Value persistence bug causing NaN after deploy/restart<br/>
+-   IMPROVED: Weekly and monthly reset logic for more robust period handling<br/>
+
 **Version 0.7.2** - January 2025<br/>
 
 -   Added: example and explanation using filepath storage<br/>